diffusr

Network diffusion algorithms in R.

References

Tong, H., Faloutsos, C., & Pan, J. Y. (2006), Fast random walk with restart and its applications.

Koehler, S., Bauer, S., Horn, D., & Robinson, P. N. (2008), Walking the interactome for prioritization of candidate disease genes. The American Journal of Human Genetics

Bonacich, P. (1987), Power and centrality: A family of measures. American Journal of Sociology

Leiserson, M. D., Vandin, F., Wu, H. T., Dobson, J. R., Eldridge, J. V., Thomas, J. L., ... & Lawrence, M. S. (2015), Pan-cancer network analysis identifies combinations of rare somatic mutations across pathways and protein complexes. Nature genetics

https://en.wikipedia.org/wiki/Laplacian_matrix

https://en.wikipedia.org/wiki/Heat_equation

Author

Simon Dirmeier simon.dirmeier@gmx.de